Move all types to types.go
