+++ api/cluster_alert.go
@@ -3,8 +3,6 @@ package api
 import (
 	"fmt"
 	"time"
-
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 const (
@@ -12,55 +10,6 @@ const (
 	ResourceNameClusterAlert = "cluster-alert"
 	ResourceTypeClusterAlert = "clusteralerts"
 )
-
-// ClusterAlert types for appscode.
-type ClusterAlert struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard object's metadata.
-	// More info: http://releases.k8s.io/release-1.2/docs/devel/api-conventions.md#metadata
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-
-	// Spec is the desired state of the ClusterAlert.
-	// More info: http://releases.k8s.io/release-1.2/docs/devel/api-conventions.md#spec-and-status
-	Spec ClusterAlertSpec `json:"spec,omitempty"`
-
-	// Status is the current state of the ClusterAlert.
-	// More info: http://releases.k8s.io/release-1.2/docs/devel/api-conventions.md#spec-and-status
-	// Status AlertStatus `json:"status,omitempty"`
-}
-
-// ClusterAlertList is a collection of ClusterAlert.
-type ClusterAlertList struct {
-	metav1.TypeMeta `json:",inline"`
-	// Standard object's metadata.
-	// More info: http://releases.k8s.io/release-1.2/docs/devel/api-conventions.md#metadata
-	metav1.ListMeta `json:"metadata,omitempty"`
-
-	// Items is the list of ClusterAlert.
-	Items []ClusterAlert `json:"items"`
-}
-
-// ClusterAlertSpec describes the ClusterAlert the user wishes to create.
-type ClusterAlertSpec struct {
-	// Icinga CheckCommand name
-	Check CheckCluster `json:"check,omitempty"`
-
-	// How frequently Icinga Service will be checked
-	CheckInterval metav1.Duration `json:"checkInterval,omitempty"`
-
-	// How frequently notifications will be send
-	AlertInterval metav1.Duration `json:"alertInterval,omitempty"`
-
-	// Secret containing notifier credentials
-	NotifierSecretName string `json:"notifierSecretName,omitempty"`
-
-	// NotifierParams contains information to send notifications for Incident
-	// State, UserUid, Method
-	Receivers []Receiver `json:"receivers,omitempty"`
-
-	// Vars contains Icinga Service variables to be used in CheckCommand
-	Vars map[string]interface{} `json:"vars,omitempty"`
-}
 
 var _ Alert = &ClusterAlert{}
